changed awaiting payload

+++ src/timer.c
@@ -50,7 +50,7 @@ void delay(uint16_t time){
 }
 
 uint16_t durationPOT(void){
-    uint32_t result = ADC0.RESULT;
+    uint16_t result = ADC0.RESULT;
     uint16_t duration = (result * ADC8bit) + ((int16_t)(result * ADC8bit) >> 8) + 250;
     return duration;
 }
